requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
jinja2>=3.1,<4.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0
passlib>=1.7,<2.0
email-validator>=2.1,<3.0
itsdangerous>=2.2,<3.0
supabase>=2.10,<3.0
